+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for Fontra
 
+## 2025-03-16
+
+- Fixed several bugs in the designspace backend related to editing font sources. [Issue 2040](https://github.com/googlefonts/fontra/issues/2040) [PR 2091](https://github.com/googlefonts/fontra/pull/2091)
+
 ## 2025-03-14
 
 New features: